67 WIP for create / edit game form refactor.
