+++ client/src/components/Games/queries.js
@@ -21,6 +21,7 @@ fragment gameFields on Game {
     publicId
   }
   label {
+    id
     displayName
     shortName
   }
